完成表面上的oo化
